Move custom files
